feat(auth/payments): adds user verification in payment management

+++ src/use-cases/Payment/DeletePaymentUseCase.ts
@@ -1,3 +1,4 @@
+import { ERRORS_MESSAGES } from "../../constants/Errors";
 import { PaymentRepository } from "../../repositories/PaymentRepository";
 
 export class DeletePaymentUseCase {
@@ -6,7 +7,13 @@ export class DeletePaymentUseCase {
     private paymentRepository: PaymentRepository
   ) {}
 
-  async execute(paymentUuid: string) {
+  async execute(userId: number, paymentUuid: string) {
+    const payment = await this.paymentRepository.getByUuid(paymentUuid);
+
+    if (payment?.Customer?.userId !== userId) {
+      throw new Error(ERRORS_MESSAGES.UNAUTHORIZED);
+    }
+
     await this.paymentRepository.delete(paymentUuid);
   }
 }
